# Performance Profiler: Remove Generated with AI badge as it confuses users