--- a/client/performance-profiler/components/insights-section/index.tsx
+++ b/client/performance-profiler/components/insights-section/index.tsx
@@ -1,7 +1,4 @@
 import { SelectDropdown } from '@automattic/components';
-import { useDesktopBreakpoint } from '@automattic/viewport-react';
-import styled from '@emotion/styled';
-import clsx from 'clsx';
 import { useTranslate } from 'i18n-calypso';
 import { ForwardedRef, forwardRef, useCallback, useEffect, useState } from 'react';
 import {
@@ -29,36 +26,9 @@ type InsightsSectionProps = {
 	onRecommendationsFilterChange?: ( filter: string ) => void;
 };
 
-const AIBadge = styled.span`
-	padding: 0 8px;
-	margin-left: 8px;
-	margin-top: 2px;
-	width: fit-content;
-	height: fit-content;
-	border-radius: 4px;
-	float: right;
-	font-size: 12px;
-	line-height: 20px;
-	color: var( --studio-gray-100 );
-	background: linear-gradient(
-			0deg,
-			rgba( 255, 255, 255, 0.95 ) 0%,
-			rgba( 255, 255, 255, 0.95 ) 100%
-		),
-		linear-gradient( 90deg, #4458e4 0%, #069e08 100% );
-
-	&.is-mobile {
-		float: none;
-		display: block;
-		margin-left: 0;
-		margin-top: 8px;
-	}
-`;
-
 export const InsightsSection = forwardRef(
 	( props: InsightsSectionProps, ref: ForwardedRef< HTMLDivElement > ) => {
 		const translate = useTranslate();
-		const isMobile = ! useDesktopBreakpoint();
 		const { audits, fullPageScreenshot, isWpcom, hash, filter, onRecommendationsFilterChange } =
 			props;
 		const [ selectedFilter, setSelectedFilter ] = useState( filter ?? 'all' );
@@ -93,12 +63,7 @@ export const InsightsSection = forwardRef(
 			<div className="performance-profiler-insights-section" ref={ ref }>
 				<div className="header">
 					<div>
-						<h2 className="title">
-							{ translate( 'Personalized Recommendations' ) }
-							<AIBadge className={ clsx( { 'is-mobile': isMobile } ) }>
-								{ translate( 'Generated with AI' ) }
-							</AIBadge>
-						</h2>
+						<h2 className="title">{ translate( 'Personalized Recommendations' ) }</h2>
 						<p className="subtitle">
 							{ getSubtitleText( selectedFilter, filteredAudits.length, translate ) }
 						</p>